ml/requirements.txt
Flask==3.1.3
scikit-learn==1.6.1
numpy==2.2.0
pandas==2.3.1
supabase==2.23.0
python-dotenv==1.1.1
stumpy==1.14.1
matplotlib==3.10.3
scipy==1.15.3
requests==2.31.0
